fix(mobile): show jump-to-latest in thread views
Fixes #5650
