+++ mobile/lib/features/channels/thread_detail_page.dart
@@ -25,6 +25,7 @@ import 'date_formatters.dart';
 import 'day_divider.dart';
 import '../profile/user_profile_sheet.dart';
 import 'initial_thread_tail_settle.dart';
+import 'jump_to_latest_button.dart';
 import 'laid_out_viewport.dart';
 import 'message_actions.dart';
 import 'message_long_press_region.dart';
@@ -37,6 +38,7 @@ import 'small_avatar.dart';
 import 'timeline_message.dart';
 
 part 'thread_detail_helpers.dart';
+part 'thread_detail_nested_summary.dart';
 
 /// Full-screen thread detail page.
 ///
@@ -111,6 +113,7 @@ class ThreadDetailPage extends HookConsumerWidget {
     final didJumpToInitialMessage = useRef(false);
     final followsThreadTail = useRef(false);
     final userOptedOutOfTailFollow = useRef(false);
+    final isAtLatest = useState(true);
     final tailIntent = useMemoized(_ThreadTailIntent.new);
     final pendingTailAlignment = useRef<double?>(null);
     const headIndex = 0;
@@ -132,8 +135,14 @@ class ThreadDetailPage extends HookConsumerWidget {
 
     useEffect(() {
       void onPositionsChanged() {
-        if (!userOptedOutOfTailFollow.value && threadTailIsVisible()) {
+        final tailVisible = threadTailIsVisible();
+        if (!userOptedOutOfTailFollow.value && tailVisible) {
           followsThreadTail.value = true;
+        }
+        if (tailVisible) {
+          if (!isAtLatest.value) isAtLatest.value = true;
+        } else if (userOptedOutOfTailFollow.value && isAtLatest.value) {
+          isAtLatest.value = false;
         }
       }
 
@@ -165,6 +174,7 @@ class ThreadDetailPage extends HookConsumerWidget {
         action: () {
           tailIntent.detach();
           followsThreadTail.value = false;
+          isAtLatest.value = false;
           pendingTailAlignment.value = null;
           itemScrollController.jumpTo(index: targetIndex, alignment: 0.35);
         },
@@ -228,6 +238,11 @@ class ThreadDetailPage extends HookConsumerWidget {
           }
         },
       );
+    }
+
+    void scrollToLatest() {
+      if (!itemScrollController.isAttached || tailIntent.isDragging) return;
+      queueTailRealignment(allowIdleDetached: true, restoreFollow: true);
     }
 
     useEffect(() {
@@ -559,6 +574,21 @@ class ThreadDetailPage extends HookConsumerWidget {
                 _ThreadTypingIndicator(entries: threadTyping, animated: false),
             ],
           ),
+          if (!isAtLatest.value)
+            Positioned(
+              left: 0,
+              right: 0,
+              bottom:
+                  (isMember && !isArchived ? composerDockHeight.value : 0) +
+                  Grid.xs,
+              child: Center(
+                child: JumpToLatestButton(
+                  key: const ValueKey('thread-jump-to-latest'),
+                  surfaceKey: const ValueKey('thread-jump-to-latest-surface'),
+                  onPressed: scrollToLatest,
+                ),
+              ),
+            ),
           if (isMember && !isArchived)
             Align(
               alignment: Alignment.bottomCenter,
@@ -593,118 +623,6 @@ class ThreadDetailPage extends HookConsumerWidget {
               ),
             ),
         ],
-      ),
-    );
-  }
-}
-
-/// Tappable summary row shown below a reply that itself has replies.
-/// Pushes a new [ThreadDetailPage] for the nested thread.
-class _NestedThreadSummaryRow extends ConsumerWidget {
-  final ThreadSummary summary;
-  final TimelineMessage replyMessage;
-  final List<TimelineMessage> allMessages;
-  final String channelId;
-  final String? currentPubkey;
-  final bool isMember;
-  final bool isArchived;
-
-  const _NestedThreadSummaryRow({
-    required this.summary,
-    required this.replyMessage,
-    required this.allMessages,
-    required this.channelId,
-    required this.currentPubkey,
-    required this.isMember,
-    required this.isArchived,
-  });
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final userCache = ref.watch(userCacheProvider);
-
-    return GestureDetector(
-      onTap: () {
-        Navigator.of(context).push(
-          MaterialPageRoute<void>(
-            builder: (_) => ThreadDetailPage(
-              threadHead: replyMessage,
-              allMessages: allMessages,
-              channelId: channelId,
-              currentPubkey: currentPubkey,
-              isMember: isMember,
-              isArchived: isArchived,
-            ),
-          ),
-        );
-      },
-      child: Padding(
-        key: ValueKey('nested-thread-summary-${replyMessage.id}'),
-        padding: const EdgeInsets.only(
-          left: messageAvatarSize + messageAvatarContentGap,
-          top: Grid.half,
-          bottom: Grid.xs,
-        ),
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            // Stacked participant avatars.
-            SizedBox(
-              width:
-                  32.0 +
-                  (summary.participantPubkeys.length - 1).clamp(0, 2) * 20.0,
-              height: 32,
-              child: Stack(
-                children: [
-                  for (var i = 0; i < summary.participantPubkeys.length; i++)
-                    Positioned(
-                      left: i * 20.0,
-                      child: SmallAvatar(
-                        pubkey: summary.participantPubkeys[i],
-                        userCache: userCache,
-                        size: 32,
-                      ),
-                    ),
-                ],
-              ),
-            ),
-            const SizedBox(width: Grid.xxs),
-            Flexible(
-              child: Text.rich(
-                TextSpan(
-                  children: [
-                    TextSpan(
-                      text:
-                          '${summary.replyCount} ${summary.replyCount == 1 ? 'reply' : 'replies'}',
-                      style: replyPreviewTextStyle.copyWith(
-                        color: context.colors.primary,
-                      ),
-                    ),
-                    if (summary.lastReplyAt case final lastReplyAt?) ...[
-                      TextSpan(
-                        text: ' · ',
-                        style: replyPreviewTextStyle.copyWith(
-                          color: context.colors.onSurfaceVariant.withValues(
-                            alpha: 0.5,
-                          ),
-                        ),
-                      ),
-                      TextSpan(
-                        text:
-                            'last reply ${formatThreadSummaryLastReplyTime(lastReplyAt)}',
-                        style: replyPreviewTextStyle.copyWith(
-                          color: context.colors.onSurfaceVariant,
-                        ),
-                      ),
-                    ],
-                  ],
-                ),
-                maxLines: 2,
-                overflow: TextOverflow.ellipsis,
-              ),
-            ),
-          ],
-        ),
       ),
     );
   }
